+++ src/engine/professor-intent-handler.js
@@ -1,3 +1,3 @@
 'use strict';
-const { requestedProfessorFields, professorIntentLabel, formatProfessorFieldResponse } = require('../professor-card-response');
-module.exports = { requestedProfessorFields, professorIntentLabel, formatProfessorFieldResponse };
+const { requestedProfessorFields, professorIntentLabel, formatProfessorFieldResponse, isProfessorPrivatePhoneRequest, formatProfessorPhonePrivacyResponse } = require('../professor-card-response');
+module.exports = { requestedProfessorFields, professorIntentLabel, formatProfessorFieldResponse, isProfessorPrivatePhoneRequest, formatProfessorPhonePrivacyResponse };